Cardano.ChainProducer

Description

Documentation

class MonadChainProducer m where

Methods

nextBlocks

Arguments

:: Natural
Number of blocks to retrieve

-> SlotId
Starting point

-> ExceptT ErrGetNextBlocks m [Block]

Get some blocks from the chain producer.

This may retrieve less than the requested number of blocks. It might return no blocks at all.

newtype ErrGetNextBlocks

The things that can go wrong when retrieving blocks.

Constructors

GetNextBlocksError String