Cardano.ChainProducer.RustHttpBridge.Client

Description

An API client for the Cardano HTTP Bridge.

Documentation

getBlockByHash :: NetworkName -> Hash Blake2b_256 BlockHeader -> ClientM Block

Retrieve a block identified by the unique hash of its header.

getEpochById :: NetworkName -> EpochId -> ClientM [Block]

Retrieve all the blocks for the epoch identified by the given integer ID.

getTipBlockHeader :: NetworkName -> ClientM (WithHash Blake2b_256 BlockHeader)

Retrieve the header of the latest known block.