Cardano.Environment

Description

This module contains static configuration parameters. Rather than providing and carrying around a configuration file through the application, we resolve configuration data at runtime using the available environment.

This gives us a flexible and portable approach to software configuration, and remove some pain from the development perspective. Prior to starting, the wallet is expected to have a few configuration parameter available. One may rely on a `.env` file to bundle configuration settings together for a given target environment.

Networking

data Network

Available network options. Local means a local cluster running on the host machine.

Constructors

Mainnet

Testnet

Staging

Local

network :: Network

Get the current target Network from the Environment.

Throws a runtime exception is the ENV var isn't set or, is invalid.

newtype ProtocolMagic

Constructors

ProtocolMagic Int32

protocolMagic :: Network -> ProtocolMagic

Get the ProtocolMagic corresponding to a given Network.

Note that the ProtocolMagic for Local and Testnet are the same.

Internals

data ErrMissingOrInvalidEnvVar

Fatal exception thrown when a required ENV var is missing upon start-up.

Constructors

ErrMissingOrInvalidEnvVar

Fields

name :: String

command :: String

additionalContext :: Maybe (String, TextDecodingError)

unsafeLookupEnv :: FromText a => String -> a

Lookup the environment for a given variable