Cardano.NetworkLayer.HttpBridge

Description

This module contains the necessary logic to talk to implement the network layer using the cardano-http-bridge as a chain producer.

Documentation

newtype RustBackend a

Constructors

RustBackend

Fields

runRB :: ReaderT (HttpBridge IO) IO a

runRustBackend :: HttpBridge IO -> RustBackend a -> IO a

getNetwork :: RustBackend (HttpBridge IO)

newtype ErrGetNextBlocks

The things that can go wrong when retrieving blocks.

Constructors

GetNextBlocksError String

nextBlocks

Arguments

:: Word64
Number of blocks to retrieve

-> SlotId
Starting point

-> ExceptT ErrGetNextBlocks RustBackend [Block]

epochRange

Arguments

:: Word64
Number of slots

-> SlotId
Start point

-> SlotId
Latest block available

-> [Word64]

Calculates which epochs to fetch, given a number of slots, and the start point. It takes into account the latest block available, and that the most recent epoch is never available in a pack file.

getEpochs :: Monad m => HttpBridge m -> [Word64] -> ExceptT HttpBridgeError m [[Block]]

Fetch epoch blocks until one fails.

fetchBlocksFromTip :: Monad m => HttpBridge m -> SlotId -> Hash "BlockHeader" -> ExceptT HttpBridgeError m [Block]

runHttpBridge :: ExceptT HttpBridgeError IO a -> ExceptT ErrGetNextBlocks RustBackend a

Utility functions for monadic loops

mapUntilError

Arguments

:: Monad m

=> (a -> ExceptT e m b)
Action to run

-> [a]
Elements

-> ExceptT e m [b]
Results

Apply an action to each element of a list, until an action fails, or there are no more elements. This is like mapM, except that it always succeeds and the resulting list might be smaller than the given list.

data HttpBridge m

Endpoints of the cardano-http-bridge API.

Constructors

HttpBridge

Fields

getBlock :: Hash "BlockHeader" -> ExceptT HttpBridgeError m Block

getEpoch :: Word64 -> ExceptT HttpBridgeError m [Block]

getNetworkTip :: ExceptT HttpBridgeError m (Hash "BlockHeader", BlockHeader)

newtype HttpBridgeError

Constructors

HttpBridgeError String

getBlockByHash :: NetworkName -> Hash Blake2b_256 (ApiT BlockHeader) -> ClientM (ApiT Block)

Retrieve a block identified by the unique hash of its header.

getEpochById :: NetworkName -> EpochIndex -> ClientM [ApiT Block]

Retrieve all the blocks for the epoch identified by the given integer ID.

getTipBlockHeader :: NetworkName -> ClientM (WithHash Blake2b_256 (ApiT BlockHeader))

Retrieve the header of the latest known block.

mkHttpBridge :: Manager -> BaseUrl -> NetworkName -> HttpBridge IO

Construct a new network layer

blockHeaderHash :: WithHash algorithm (ApiT BlockHeader) -> (Hash "BlockHeader", BlockHeader)

hashToApi :: HashAlgorithm a => Hash h -> Maybe (Hash a b)

hashToApi' :: (Monad m, HashAlgorithm algorithm) => Hash a -> ExceptT HttpBridgeError m (Hash algorithm b)

Converts a Hash to the Digest type that the Api module requires.