Cardano.Wallet

Description

Here we find the "business logic" to manage a Cardano wallet. This is a direct implementation of the model from the Formal Specification for a Cardano Wallet Note that, this module is purposedly agnostic to how blocks are retrieved or how various types are serialized.

This is really about how the wallet keep track of its internal state, and its UTxO (where the relationship is defined via the IsOurs abstraction to allow this core code to be used with any sort of derivation scheme).

All those functions are pure and there's no reason to shove in any sort of side-effects in here :)

Wallet

data Wallet s

An opaque wallet type, see initWallet and applyBlock to construct and update wallets.

Internally, this keeps track or a few things including:

UTxOs

Pending transaction

TODO: Transaction history

TODO: Known & used addresses

initWallet :: (IsOurs s, Semigroup s, NFData s, Show s) => s -> Wallet s

Create an empty wallet from an initial state

currentTip :: Wallet s -> SlotId

Get the wallet current tip

applyBlock :: Block -> NonEmpty (Wallet s) -> NonEmpty (Wallet s)

Apply Block is the only way to make the wallet evolve.

availableBalance :: Wallet s -> Integer

Available balance = balance . availableUTxO

totalBalance :: Wallet s -> Integer

Total balance = balance . totalUTxO

totalUTxO :: Wallet s -> UTxO

Total UTxO = availableUTxO <> "pending UTxO"

availableUTxO :: Wallet s -> UTxO

Available UTxO = UTxO that aren't part of pending txs

txOutsOurs :: forall s. IsOurs s => Set Tx -> s -> (Set TxOut, s)

Return all transaction outputs that are ours. This plays well within a State monad.

myFunction :: Block -> State s Result
myFunction b = do
   ours <- state $ txOutsOurs (transaction b)
   return $ someComputation ours

utxoFromTx :: Tx -> UTxO

Construct a UTxO corresponding to a given transaction. It is important for the transaction outputs to be ordered correctly, since they become available inputs for the subsequent blocks.

Wallet Metadata

data WalletMetadata

Constructors

WalletMetadata

Fields

id :: !WalletId

name :: !WalletName

addressPoolGap :: !AddressPoolGap

passphraseInfo :: !WalletPassphraseInfo

status :: !WalletStatus

delegation :: !WalletDelegation

newtype WalletId

Constructors

WalletId Text

newtype WalletName

Constructors

WalletName

Fields

getWalletName :: Text

newtype WalletTimestamp

Constructors

WalletTimestamp Microsecond

data WalletStatus

Constructors

Ready

Restoring

data WalletDelegation

Constructors

Delegated

NotDelegated

newtype WalletPassphraseInfo

Constructors

WalletPassphraseInfo

Fields

lastUpdated :: WalletTimestamp