Cardano.Wallet.AddressDiscovery

Description

This module contains primitives necessary to perform address discovery. So far, we're only interested in address following a sequential derivation scheme as specified in BIP-0044. Later, we may introduce backward compatibility with random address scheme from the legacy Cardano wallets.

Sequential Derivation

Address Pool Gap

data AddressPoolGap

Maximum number of consecutive undiscovered addresses allowed

newtype MkAddressPoolGapError

Possible errors when casting to an AddressPoolGap

Constructors

ErrGapOutOfRange Word8

defaultAddressPoolGap :: AddressPoolGap

A default AddressPoolGap, as suggested in BIP-0044

mkAddressPoolGap :: Word8 -> Either MkAddressPoolGapError AddressPoolGap

Smart constructor for AddressPoolGap

Address Pool

data AddressPool

An AddressPool which keeps track of sequential addresses within a given Account and change chain. See mkAddressPool to create a new or existing pool:

>>> mkAddressPool xpub gap changeChain mempty
AddressPool { }

gap :: AddressPool -> AddressPoolGap

The actual gap for the pool. This can't change for a given pool.

addresses :: AddressPool -> [Address]

Get all addresses in the pool, sorted from the first address discovered, up until the next one.

In practice, we always have:

mkAddressPool key g cc (addresses pool) == pool

changeChain :: AddressPool -> ChangeChain

Whether this pool tracks addrs on the internal or external chain

accountPubKey :: AddressPool -> Key AccountK XPub

Corresponding key for the pool (a pool is tied to only one account)

mkAddressPool :: Key AccountK XPub -> AddressPoolGap -> ChangeChain -> [Address] -> AddressPool

Create a new Address pool from a list of addresses. Note that, the list is expected to be ordered in sequence (first indexes, first in the list).

The pool will grow from the start if less than g :: AddressPoolGap are given, such that, there are always g undiscovered addresses in the pool.

lookupAddress :: Address -> AddressPool -> (Maybe (Index Soft AddressK), AddressPool)

Lookup an address in the pool. When we find an address in a pool, the pool may be amended if the address was discovered near the edge. It is also possible that the pool is not amended at all - this happens in the case that an address is discovered far from the edge.