Cardano.Wallet.Api.Types

API Types

newtype Amount

Constructors

Amount Natural

newtype Percentage

Constructors

Percentage Int

newtype PoolId

Constructors

PoolId

Fields

_uuid :: UUID

data Wallet

Constructors

Wallet

Fields

_id :: !WalletId

_addressPoolGap :: !(ApiT AddressPoolGap)

_balance :: !WalletBalance

_delegation :: !WalletDelegation

_name :: !(ApiT WalletName)

_passphrase :: !WalletPassphraseInfo

_state :: !WalletState

data WalletBalance

Constructors

WalletBalance

Fields

_available :: !Amount

_total :: !Amount

data WalletDelegation

Wallet Delegation representation, can be serialized to and from JSON as follows:

>>> Aeson.encode NotDelegating
{"status":"not_delegating"}

>>> Aeson.encode $ Delegating poolId
{"status":"delegating","target": "27522fe5-262e-42a5-8ccb-cef884ea2ba0"}

Constructors

NotDelegating

Delegating !PoolId

newtype WalletId

Constructors

WalletId

Fields

_uuid :: UUID

newtype WalletPassphraseInfo

Constructors

WalletPassphraseInfo

Fields

_lastUpdatedAt :: UTCTime

data WalletState

A Wallet State representation in the API. It can be serialized to and from JSON as follows:

>>> Aeson.encode Ready
{"status":"ready"}

>>> Aeson.encode $ Restoring (Percentage 14)
{"status":"restoring","progress":{"quantity":14,"unit":"percent"}}

Constructors

Ready

Restoring !Percentage

Re-Exports From Primitive Types

data AddressPoolGap

Maximum number of consecutive undiscovered addresses allowed

newtype WalletName

Constructors

WalletName

Fields

getWalletName :: Text

Polymorphic Types

newtype ApiT a

Constructors

ApiT

Fields

getApiT :: a

newtype MeasuredIn (u :: Symbol) a

Represents a value that has an associated unit of measure, based on some underlying type.

Constructors

MeasuredIn a