Cardano.Wallet.Binary

Description

These are (partial) CBOR decoders for blocks and block headers. Note that we ignore most of the block's and header's content and only retrieve the pieces of information relevant to us, wallet (we do assume a trusted node and therefore, we needn't to care about verifying signatures and blocks themselves).

Documentation

decodeBlock :: Decoder s Block

decodeBlockHeader :: Decoder s BlockHeader