Cardano.Wallet.Binary.Helpers

Description

Some extra helpers on top of the CBOR library to help writing decoders.

Documentation

inspectNextToken :: Decoder s ()

Inspect the next token that has to be decoded and print it to the console as a trace. Useful for debugging Decoders. Example:

myDecoder :: CBOR.Decoder s MyType myDecoder = do a <- CBOR.decodeWord64 inspectNextToken [...]

decodeList :: forall s a. Decoder s a -> Decoder s [a]

Decode an list of known length. Very similar to decodeListIndef.

Example:

myDecoder :: CBOR.Decoder s [MyType] myDecoder = decodeList decodeOne where decodeOne :: CBOR.Decoder s MyType

decodeListIndef :: forall s a. Decoder s a -> Decoder s [a]

Decode an arbitrary long list. CBOR introduce a "break" character to mark the end of the list, so we simply decode each item until we encounter a break character.

Example:

myDecoder :: CBOR.Decoder s [MyType] myDecoder = decodeListIndef decodeOne where decodeOne :: CBOR.Decoder s MyType