Cardano.Wallet.BlockSyncer

Description

This module contains the ticking function that is responsible for invoking block acquisition functionality and executing it in periodic fashion.

Known limitations: - Blocks are produced by the network layer. They are not expected to produce any duplicates and to rollback.

Documentation

tick

Arguments

:: MonadIO m

=> (st -> m ([b], st))
A way to get a new elements

-> (b -> m ())
Action to be taken on new elements

-> Millisecond
tick time

-> st

-> m ()

Every interval delay, fetches some data from a given source, and call an action for each elements retrieved.

listen :: forall e0 e1. Show e0 => NetworkLayer IO e0 e1 -> (Block -> IO ()) -> IO ()

Retrieve blocks from a chain producer and execute some given action for each block.