Cardano.Wallet.CoinSelection

Description

Provides the API of Coin Selection algorithm and Fee related functionality. For more information refer to: https://iohk.io/blog/self-organisation-in-coin-selection/

Documentation

newtype CoinSelectionOptions

Constructors

CoinSelectionOptions

Fields

maximumNumberOfInputs :: Word64

data CoinSelectionError

Constructors

NotEnoughMoney Word64 Word64
UTxO exhausted during input selection We record the balance of the UTxO as well as the size of the payment we tried to make.

UtxoNotEnoughFragmented Word64 Word64
UTxO is not enough fragmented for the number of transaction outputs We record the number of UTxO entries as well as the number of the outputs of the transaction.

MaximumInputsReached Word64
When trying to construct a transaction, the max number of allowed inputs was reached.

data CoinSelection

Constructors

CoinSelection

Fields

inputs :: [(TxIn, TxOut)]

Picked inputs

outputs :: [TxOut]

Picked outputs

change :: [Coin]

Resulting changes