Cardano.Wallet.Mnemonic

Description

This module provides mnemonic (backup phrase) creation, and conversion of a mnemonic to seed for wallet restoration.

Types

data Mnemonic (mw :: Nat)

A backup-phrase in the form of a non-empty of Mnemonic words Constructor isn't exposed.

data Entropy (n :: Nat)

type family EntropySize (n :: Nat) :: Nat where ...

Equations

EntropySize 9 = 96

EntropySize 12 = 128

EntropySize 15 = 160

EntropySize 18 = 192

EntropySize 21 = 224

EntropySize 24 = 256

type family MnemonicWords (n :: Nat) :: Nat where ...

Equations

MnemonicWords 96 = 9

MnemonicWords 128 = 12

MnemonicWords 160 = 15

MnemonicWords 192 = 18

MnemonicWords 224 = 21

MnemonicWords 256 = 24

Errors

data MnemonicError csz

This wraps errors from Cardano.Encoding.BIP39

Constructors

ErrMnemonicWords MnemonicWordsError
Wrong number of words in mnemonic.

ErrEntropy (EntropyError csz)
Invalid entropy length or checksum.

ErrDictionary DictionaryError
Invalid word in mnemonic.

newtype MnemonicException csz

This wraps EntropyError of Cardano.Encoding.BIP39

Constructors

UnexpectedEntropyError (EntropyError csz)
Invalid entropy length or checksum

data EntropyError (csz :: Nat)

Constructors

ErrInvalidEntropyLength Int Int

ErrInvalidEntropyChecksum (Checksum csz) (Checksum csz)

data DictionaryError

Constructors

ErrInvalidDictionaryWord String

data MnemonicWordsError

Constructors

ErrWrongNumberOfWords Int Int

Creating Mnemonic (resp. Entropy)

mkEntropy :: forall n csz. (ValidEntropySize n, ValidChecksumSize n csz) => ByteString -> Either (EntropyError csz) (Entropy n)

Smart-constructor for the Entropy

mkMnemonic :: forall mw n csz. (ConsistentEntropy n mw csz, EntropySize mw ~ n) => [Text] -> Either (MnemonicError csz) (Mnemonic mw)

Smart-constructor for the Mnemonic

genEntropy :: forall n csz. (ValidEntropySize n, ValidChecksumSize n csz) => IO (Entropy n)

Generate Entropy of a given size using a random seed.

Example: do ent <- genEntropy :: IO (Entropy 12)

Converting from and to Mnemonic (resp. Entropy)

mnemonicToEntropy :: Mnemonic mw -> Entropy (EntropySize mw)

entropyToMnemonic :: forall mw n csz. (ValidMnemonicSentence mw, ValidEntropySize n, ValidChecksumSize n csz, n ~ EntropySize mw, mw ~ MnemonicWords n) => Entropy n -> Mnemonic mw

Convert an Entropy to a corresponding Mnemonic Sentence

entropyToBytes :: Entropy n -> ScrubbedBytes

Convert Entropy to a raw ByteString

ambiguousNatVal :: forall n. KnownNat n => Integer

mnemonicToText :: Mnemonic mw -> [Text]

Orphan instances

Eq MnemonicWordsError

Instance details

Methods

(==) :: MnemonicWordsError -> MnemonicWordsError -> Bool

(/=) :: MnemonicWordsError -> MnemonicWordsError -> Bool

Eq DictionaryError

Instance details

Methods

(==) :: DictionaryError -> DictionaryError -> Bool

(/=) :: DictionaryError -> DictionaryError -> Bool

Eq (EntropyError czs)

Instance details

Methods

(==) :: EntropyError czs -> EntropyError czs -> Bool

(/=) :: EntropyError czs -> EntropyError czs -> Bool