Cardano.Wallet.Network

Documentation

data NetworkLayer m e0 e1

Constructors

NetworkLayer

Fields

nextBlocks :: SlotId -> ExceptT e0 m [Block]

Gets some blocks from the node. It will not necessarily return all the blocks that the node has, but will receive a reasonable-sized chunk. It will never return blocks from before the given slot. It may return an empty list if the node does not have any blocks from after the starting slot.

networkTip :: ExceptT e1 m (Hash "BlockHeader", BlockHeader)

tick

Arguments

:: MonadIO m

=> (st -> m (TickResult a, st))
A way to get a new elements

-> (a -> m ())
Action to be taken on new elements

-> Millisecond
Tick time

-> st
Initial state.

-> m ()

Repeatedly fetch data from a given source function, and call an action for each element retrieved.

If the data source indicates that it has no more data at present (Sleep), then sleep for the interval delay, and then try the fetch again.

data TickResult a

The result type of the element fetch function provided to tick.

Constructors

GotChunk !a
Have a result, and there may be more available.

Sleep
There is no result available now, so wait.

listen :: forall e0 e1. Show e0 => NetworkLayer IO e0 e1 -> ([Block] -> IO ()) -> IO ()

Retrieve blocks from a chain producer and execute some given action for each block.