Cardano.Wallet.Primitive

Description

This module contains the core primitive of a Wallet. This is roughly a Haskell translation of the Formal Specification for a Cardano Wallet

It doesn't contain any particular business-logic code, but define a few primitive operations on Wallet core types as well.

Block

data Block

Constructors

Block

Fields

header :: !BlockHeader

transactions :: !(Set Tx)

data BlockHeader

Constructors

BlockHeader

Fields

epochIndex :: !EpochIndex

slotNumber :: !LocalSlotIndex

prevBlockHash :: !(Hash "BlockHeader")

Tx

data Tx

Constructors

Tx

Fields

inputs :: ![TxIn]

Order of inputs matters in the transaction representation. The transaction id is computed from the binary representation of a tx, for which inputs are serialized in a specific order.

outputs :: ![TxOut]

Order of outputs matter in the transaction representations. Outputs are used as inputs for next transactions which refer to them using their indexes. It matters also for serialization.

data TxIn

Constructors

TxIn

Fields

inputId :: !(Hash "Tx")

inputIx :: !Word32

data TxOut

Constructors

TxOut

Fields

address :: !Address

coin :: !Coin

txIns :: Set Tx -> Set TxIn

updatePending :: Block -> Set Tx -> Set Tx

Address

newtype Address

Constructors

Address

Fields

getAddress :: ByteString

class IsOurs s where

This abstraction exists to give us the ability to keep the wallet business logic agnostic to the address derivation and discovery mechanisms.

This is needed because two different address schemes lives on Cardano: - A hierarchical random scheme: rather custom made, with several flaws; this is the original and now legacy address scheme.

A hierarchical sequential scheme: a new scheme based on the BIP-0044 specification, which is better suited for our present needs.

In practice, we will need a wallet that can support both, even if not at the same time, and this little abstraction can buy us this without introducing too much overhead.

Methods

isOurs :: Address -> s -> (Bool, s)

Coin

newtype Coin

Coins are stored as Lovelace (reminder: 1 Lovelace = 1e6 ADA)

Constructors

Coin

Fields

getCoin :: Word64

isValidCoin :: Coin -> Bool

UTxO

newtype UTxO

Constructors

UTxO

Fields

getUTxO :: Map TxIn TxOut

balance :: UTxO -> Integer

excluding :: UTxO -> Set TxIn -> UTxO

isSubsetOf :: UTxO -> UTxO -> Bool

restrictedBy :: UTxO -> Set TxIn -> UTxO

restrictedTo :: UTxO -> Set TxOut -> UTxO

class Dom a where

Associated Types

type DomElem a :: *

Methods

dom :: a -> Set (DomElem a)

Generic

newtype Hash (tag :: Symbol)

Constructors

Hash

Fields

getHash :: ByteString

newtype ShowFmt a

A polymorphic wrapper type with a custom show instance to display data through Buildable instances.

Constructors

ShowFmt a