Cardano.Wallet.Primitive.Types

Description

This module contains the core primitive of a Wallet. This is roughly a Haskell translation of the Formal Specification for a Cardano Wallet

It doesn't contain any particular business-logic code, but define a few primitive operations on Wallet core types as well.

Block

data Block

Constructors

Block

Fields

header :: !BlockHeader

transactions :: ![Tx]

data BlockHeader

Constructors

BlockHeader

Fields

slotId :: SlotId

prevBlockHash :: !(Hash "BlockHeader")

Tx

data Tx

Constructors

Tx

Fields

inputs :: ![TxIn]

Order of inputs matters in the transaction representation. The transaction id is computed from the binary representation of a tx, for which inputs are serialized in a specific order.

outputs :: ![TxOut]

Order of outputs matter in the transaction representations. Outputs are used as inputs for next transactions which refer to them using their indexes. It matters also for serialization.

data TxIn

Constructors

TxIn

Fields

inputId :: !(Hash "Tx")

inputIx :: !Word32

data TxOut

Constructors

TxOut

Fields

address :: !Address

coin :: !Coin

data TxMeta

Constructors

TxMeta

Fields

status :: !TxStatus

direction :: !Direction

slotId :: !SlotId

amount :: !(Quantity "lovelace" Natural)

data Direction

The flow of funds in to or out of a wallet.

Constructors

Outgoing
Funds exit the wallet.

Incoming
Funds enter the wallet.

data TxStatus

Constructors

Pending

InLedger

Invalidated

newtype SignedTx

Wrapper around the final CBOR representation of a signed tx

Constructors

SignedTx

Fields

signedTx :: ByteString

txIns :: Set Tx -> Set TxIn

Address

class IsOurs s where

This abstraction exists to give us the ability to keep the wallet business logic agnostic to the address derivation and discovery mechanisms.

This is needed because two different address schemes lives on Cardano: - A hierarchical random scheme: rather custom made, with several flaws; this is the original and now legacy address scheme.

A hierarchical sequential scheme: a new scheme based on the BIP-0044 specification, which is better suited for our present needs.

In practice, we will need a wallet that can support both, even if not at the same time, and this little abstraction can buy us this without introducing too much overhead.

Methods

isOurs :: Address -> s -> (Bool, s)

newtype Address

Constructors

Address

Fields

getAddress :: ByteString

data AddressState

Constructors

Used

Unused

Coin

newtype Coin

Coins are stored as Lovelace (reminder: 1 Lovelace = 1e6 ADA)

Constructors

Coin

Fields

getCoin :: Word64

isValidCoin :: Coin -> Bool

UTxO

newtype UTxO

Constructors

UTxO

Fields

getUTxO :: Map TxIn TxOut

balance :: UTxO -> Natural

excluding :: UTxO -> Set TxIn -> UTxO

ins⋪ u

isSubsetOf :: UTxO -> UTxO -> Bool

a ⊆ b

restrictedBy :: UTxO -> Set TxIn -> UTxO

ins⊲ u

restrictedTo :: UTxO -> Set TxOut -> UTxO

u ⊳ outs

class Dom a where

Associated Types

type DomElem a :: *

Methods

dom :: a -> Set (DomElem a)

Slotting

data SlotId

A slot identifier is the combination of an epoch and slot.

Constructors

SlotId

Fields

epochIndex :: !Word64

slotNumber :: !Word16

Wallet Metadata

data WalletMetadata

Constructors

WalletMetadata

Fields

walletId :: !WalletId

name :: !WalletName

passphraseInfo :: !WalletPassphraseInfo

status :: !WalletState

delegation :: !(WalletDelegation PoolId)

newtype WalletId

Constructors

WalletId

Fields

getWalletId :: Digest Blake2b_160

newtype WalletName

Constructors

WalletName

Fields

getWalletName :: Text

walletNameMinLength :: Int

walletNameMaxLength :: Int

data WalletState

Constructors

Ready

Restoring !(Quantity "percent" Percentage)

data WalletDelegation poolId

Constructors

NotDelegating

Delegating !poolId

newtype WalletPassphraseInfo

Constructors

WalletPassphraseInfo

Fields

lastUpdatedAt :: UTCTime

data WalletBalance

Constructors

WalletBalance

Fields

available :: !(Quantity "lovelace" Natural)

total :: !(Quantity "lovelace" Natural)

Stake Pools

newtype PoolId

Represent stake pool identifier. Note that the internal representation is left open currently, until we figure out a better type for those.

Constructors

PoolId

Fields

getPoolId :: Text

Polymorphic

newtype Hash (tag :: Symbol)

Constructors

Hash

Fields

getHash :: ByteString

newtype ShowFmt a

A polymorphic wrapper type with a custom show instance to display data through Buildable instances.

Constructors

ShowFmt a

invariant

Arguments

:: String
A title / message to throw in case of violation

-> a

-> (a -> Bool)

-> a

Check whether an invariants holds or not.

>>> invariant "not empty" [1,2,3] (not . null)
[1, 2, 3]

>>> invariant "not empty" [] (not . null)
*** Exception: not empty