Cardano.Wallet.Slotting

Documentation

data SlotId

A slot identifier is the combination of an epoch and slot.

Constructors

SlotId

Fields

epochIndex :: !EpochIndex

slotNumber :: !LocalSlotIndex

newtype EpochIndex

Denotes a chain epoch, which contains a certain number of slots.

Constructors

EpochIndex

Fields

getEpochIndex :: Word64

newtype LocalSlotIndex

Denotes the slot number within an epoch.

Constructors

LocalSlotIndex

Fields

getLocalSlotIndex :: Word16

type SlotCount = Natural

slotsPerEpoch :: SlotCount

Hard-coded for the time being

addSlots :: SlotCount -> SlotId -> SlotId

Add a number of slots to an (Epoch, LocalSlotIndex) pair, where the number of slots can be greater than one epoch.

slotDiff :: SlotId -> SlotId -> Integer

slotDiff a b is the number of slots by which a is greater than b.

slotNext :: SlotId -> SlotId

The slot after.

slotPrev :: SlotId -> Maybe SlotId

The slot before, if there is one.

isValidSlotId :: SlotId -> Bool

Whether the epoch index and slot number are in range.