Cardano.WalletLayer

Description

| Module provides wallet layer api that is used by API layer | and uses both DB and Networking layer to realize its role | as being intermediary between the three.

Documentation

data CreateWalletError

Errors

data GetWalletError

data WalletLayer m s

Types

Constructors

WalletLayer

Fields

createWallet :: CreateWallet -> ExceptT CreateWalletError m (Wallet s)

getWallet :: WalletId -> ExceptT GetWalletError m (Wallet s)

data CreateWallet

Constructors

CreateWallet NewWallet

ImportWallet (Key AccountK XPub)

data NewWallet

Constructors

NewWallet

Fields

mnemonicSentence :: !(Mnemonic 15)

mnemonicSentencePassphrase :: !(Maybe (Mnemonic 9))

name :: !WalletName

passphrase :: !(Passphrase "encryption")

addressPoolGap :: !AddressPoolGap

newtype WalletName

Constructors

WalletName Text

newtype WalletId

Constructors

WalletId Text

newtype WalletTimestamp

Constructors

WalletTimestamp Microsecond

data WalletMode

Constructors

Ready

Restoring

data Delegation

Constructors

Delegated

NotDelegated

newtype PassphraseInfo

Constructors

PassphraseInfo

Fields

lastUpdated :: WalletTimestamp

data WalletMetadata

Constructors

WalletMetadata

Fields

id :: !WalletId

name :: !WalletName

addressPoolGap :: !AddressPoolGap

passphraseInfo :: !PassphraseInfo

state :: !WalletMode

delegation :: !Delegation