Data.Text.Class

Description

Extend the Text module with an extra abstraction to encode and decode values safely to and from Text. It's very similar to FromJSON and ToJSON from Aeson.

Documentation

class ToText a where

Defines a textual encoding for a type.

Methods

toText :: a -> Text

Encode the specified value as text.

class FromText a where

Defines a textual decoding for a type.

Methods

fromText :: Text -> Either TextDecodingError a

Decode the specified text as a value.

newtype TextDecodingError

Indicates an error that occurred while decoding from text.

Constructors

TextDecodingError

Fields

getTextDecodingError :: String