Cardano.NetworkLayer

Documentation

data NetworkLayer m e0 e1

Constructors

NetworkLayer

Fields

nextBlocks :: SlotId -> ExceptT e0 m [Block]

Gets some blocks from the node. It will not necessarily return all the blocks that the node has, but will receive a reasonable-sized chunk. It will never return blocks from before the given slot. It may return an empty list if the node does not have any blocks from after the starting slot.

networkTip :: ExceptT e1 m (Hash "BlockHeader", BlockHeader)

tick

Arguments

:: MonadIO m

=> (st -> m ([b], st))
A way to get a new elements

-> (b -> m ())
Action to be taken on new elements

-> Millisecond
tick time

-> st

-> m ()

Every interval delay, fetches some data from a given source, and call an action for each elements retrieved.

listen :: forall e0 e1. Show e0 => NetworkLayer IO e0 e1 -> (Block -> IO ()) -> IO ()

Retrieve blocks from a chain producer and execute some given action for each block.