Cardano.Wallet

Description

Provides the wallet layer functions that are used by API layer and uses both Cardano.Wallet.DB and Cardano.Wallet.Network to realize its role as being intermediary between the three.

Documentation

data WalletLayer s

Types

Constructors

WalletLayer

Fields

createWallet :: NewWallet -> ExceptT CreateWalletError IO WalletId

readWallet :: WalletId -> ExceptT ReadWalletError IO (Wallet s, WalletMetadata)

watchWallet :: WalletId -> IO ()

data NewWallet

Constructors

NewWallet

Fields

seed :: !(Passphrase "seed")

secondFactor :: !(Passphrase "generation")

name :: !WalletName

passphrase :: !(Passphrase "encryption")

gap :: !AddressPoolGap

newtype ReadWalletError

Errors occuring when fetching a wallet

Constructors

ErrReadWalletNotFound WalletId

newtype CreateWalletError

Errors occuring when creating a wallet

Constructors

ErrCreateWalletIdAlreadyExists WalletId

mkWalletLayer :: Show e0 => DBLayer IO SeqState -> NetworkLayer IO e0 e1 -> WalletLayer SeqState

Create a new instance of the wallet layer.