Cardano.Wallet

Description

Provides the wallet layer functions that are used by API layer and uses both Cardano.Wallet.DB and Cardano.Wallet.Network to realize its role as being intermediary between the three.

Interface

data WalletLayer s

Constructors

WalletLayer

Fields

createWallet :: NewWallet -> ExceptT ErrWalletAlreadyExists IO WalletId

Initialise and store a new wallet, returning its ID.

readWallet :: WalletId -> ExceptT ErrNoSuchWallet IO (Wallet s, WalletMetadata)

Retrieve the wallet state for the wallet with the given ID.

createUnsignedTx :: WalletId -> CoinSelectionOptions -> NonEmpty TxOut -> ExceptT ErrCreateUnsignedTx IO CoinSelection

Prepare a transaction and automatically select inputs from the wallet to cover the requested outputs. Note that this only run the coin selection for the given outputs. In order to construct (and sign) an actual transaction, have a look at $sel:signTx:WalletLayer.

signTx :: WalletId -> Key RootK XPrv -> Passphrase "encryption" -> CoinSelection -> ExceptT ErrSignTx IO (Tx, [TxWitness])

Produce witnesses and construct a transaction from a given selection. Requires the encryption passphrase in order to decrypt the root private key. Note that this doesn't broadcast the transaction to the network. In order to do so, have a look at $sel:submitTx:WalletLayer.

submitTx :: (Tx, [TxWitness]) -> ExceptT ErrSubmitTx IO ()

Broadcast a (signed) transaction to the network.

watchWallet :: forall a. WalletId -> IO a

Consume blocks from the node as they arrive, and apply them to the wallet. This function never returns, but may throw an exception.

processWallet :: (SlotId -> IO ()) -> WalletId -> IO ()

Consume the entire available chain, applying block to the given wallet, and stop when finished. This function is intended for benchmarking.

data NewWallet

Constructors

NewWallet

Fields

seed :: !(Passphrase "seed")

secondFactor :: !(Passphrase "generation")

name :: !WalletName

passphrase :: !(Passphrase "encryption")

gap :: !AddressPoolGap

Errors

newtype ErrNoSuchWallet

Can't perform given operation because there's no wallet

Constructors

ErrNoSuchWallet WalletId

newtype ErrWalletAlreadyExists

Forbidden operation was executed on an already existing wallet

Constructors

ErrWalletAlreadyExists WalletId

Construction

mkWalletLayer :: DBLayer IO SeqState -> NetworkLayer IO -> WalletLayer SeqState

Create a new instance of the wallet layer.

Helpers

unsafeRunExceptT :: (MonadFail m, Show e) => ExceptT e m a -> m a

Run an ExcepT and throws the error if any. This makes sense only if called after checking for an invariant or, after ensuring that preconditions for meeting the underlying error have been discarded.