Cardano.Wallet.Api.Types

Description

API type representations of various types. We define here pretty much all our user-facing types that are mostly composed with internal / primitive types.

This module also define required API instances (JSON, HttpApiData...) for all those types, making sure to match the specification document:

Wallet API Specification

API Types

data Address

Constructors

Address

Fields

_id :: !(ApiT Address)

_state :: !(ApiT AddressState)

data Wallet

Constructors

Wallet

Fields

_id :: !(ApiT WalletId)

_addressPoolGap :: !(ApiT AddressPoolGap)

_balance :: !(ApiT WalletBalance)

_delegation :: !(ApiT (WalletDelegation (ApiT PoolId)))

_name :: !(ApiT WalletName)

_passphrase :: !(ApiT WalletPassphraseInfo)

_state :: !(ApiT WalletState)

data WalletBalance

Constructors

WalletBalance

Fields

_available :: !(Quantity "lovelace" Natural)

_total :: !(Quantity "lovelace" Natural)

data WalletPostData

Constructors

WalletPostData

Fields

_addressPoolGap :: !(Maybe (ApiT AddressPoolGap))

_mnemonicSentence :: !(ApiMnemonicT '[15, 18, 21, 24] "seed")

_mnemonicSecondFactor :: !(Maybe (ApiMnemonicT '[9, 12] "generation"))

_name :: !(ApiT WalletName)

_passphrase :: !(ApiT (Passphrase "encryption"))

newtype WalletPutData

Constructors

WalletPutData

Fields

_name :: Maybe (ApiT WalletName)

data WalletPutPassphraseData

Constructors

WalletPutPassphraseData

Fields

_oldPassphrase :: !(ApiT (Passphrase "encryption"))

_newPassphrase :: !(ApiT (Passphrase "encryption"))

Re-Export From Primitives

data AddressState

Constructors

Used

Unused

newtype PoolId

Constructors

PoolId

Fields

getPoolId :: Text

data WalletDelegation poolId

Constructors

NotDelegating

Delegating !poolId

newtype WalletId

Constructors

WalletId UUID

newtype WalletName

Constructors

WalletName

Fields

getWalletName :: Text

newtype WalletPassphraseInfo

Constructors

WalletPassphraseInfo

Fields

lastUpdatedAt :: UTCTime

data WalletState

Constructors

Ready

Restoring !(Quantity "percent" Percentage)

data AddressPoolGap

Maximum number of consecutive undiscovered addresses allowed

newtype Passphrase (goal :: Symbol)

An encapsulated passphrase. The inner format is free, but the wrapper helps readability in function signatures.

Constructors

Passphrase ScrubbedBytes

Limits

passphraseMinLength :: Int

passphraseMaxLength :: Int

Polymorphic Types

newtype ApiT a

Polymorphic wrapper type to put around primitive types and, 3rd party lib types to avoid defining orphan instances and/or, undesirable instances on primitive types. It helps to keep a nice separation of concerns between the API layer and other modules.

Constructors

ApiT

Fields

getApiT :: a

newtype ApiMnemonicT (sizes :: [Nat]) (purpose :: Symbol)

Representation of mnemonics at the API-level, using a polymorphic type in the lengths of mnemonics that are supported (and an underlying purpose). In practice, mnemonics correspond to passphrases or seeds, and although they're nice to manipulate as mnemonics from a user-perspective, carrying around a list of words doesn't really make sense for the business logic, which prefers manipulating scrubbed bytes directly.

data MyWallet
    { mnemonic :: ApiMnemonicT '[15,18,21,24] "root-seed"
    }

Note that the given Nats **have** to be valid mnemonic sizes, otherwise the underlying code won't even compile, with not-so-friendly error messages.

Also, the internal representation holds a [Text] which contains the list of mnemonic words that was parsed. This is only to be able to implement the ToJSON instances and roundtrip, which is a very dubious argument. In practice, we'll NEVER peek at the mnemonic, output them and whatnot.

Constructors

ApiMnemonicT (Passphrase purpose, [Text])