Cardano.Wallet.Api.Types

Description

API type representations of various types. We define here pretty much all our user-facing types that are mostly composed with internal / primitive types.

This module also define required API instances (JSON, HttpApiData...) for all those types, making sure to match the specification document:

Wallet API Specification

API Types

data ApiAddress

Constructors

ApiAddress

Fields

id :: !(ApiT Address)

state :: !(ApiT AddressState)

data ApiWallet

Constructors

ApiWallet

Fields

id :: !(ApiT WalletId)

addressPoolGap :: !(ApiT AddressPoolGap)

balance :: !(ApiT WalletBalance)

delegation :: !(ApiT (WalletDelegation (ApiT PoolId)))

name :: !(ApiT WalletName)

passphrase :: !(ApiT WalletPassphraseInfo)

state :: !(ApiT WalletState)

data WalletBalance

Constructors

WalletBalance

Fields

available :: !(Quantity "lovelace" Natural)

total :: !(Quantity "lovelace" Natural)

data WalletPostData

Constructors

WalletPostData

Fields

addressPoolGap :: !(Maybe (ApiT AddressPoolGap))

mnemonicSentence :: !(ApiMnemonicT '[15, 18, 21, 24] "seed")

mnemonicSecondFactor :: !(Maybe (ApiMnemonicT '[9, 12] "generation"))

name :: !(ApiT WalletName)

passphrase :: !(ApiT (Passphrase "encryption"))

newtype WalletPutData

Constructors

WalletPutData

Fields

name :: Maybe (ApiT WalletName)

data WalletPutPassphraseData

Constructors

WalletPutPassphraseData

Fields

oldPassphrase :: !(ApiT (Passphrase "encryption"))

newPassphrase :: !(ApiT (Passphrase "encryption"))

Encoding & Decoding

newtype DecodeApiAddressError

Constructors

DecodeApiAddressError String

decodeApiAddress :: Text -> Either DecodeApiAddressError (ApiT Address)

Constructs an address from a Base58-encoded string.

Fails if the specified string is not Base58 encoded.

encodeApiAddress :: ApiT Address -> Text

Converts an address to a Base58-encoded string.

newtype DecodeApiEncryptionPassphraseError

Constructors

DecodeApiEncryptionPassphraseError String

decodeApiEncryptionPassphrase :: Text -> Either DecodeApiEncryptionPassphraseError (ApiT (Passphrase "encryption"))

encodeApiEncryptionPassphrase :: ApiT (Passphrase "encryption") -> Text

newtype DecodeApiWalletNameError

Constructors

DecodeApiWalletNameError String

decodeApiWalletName :: Text -> Either DecodeApiWalletNameError (ApiT WalletName)

encodeApiWalletName :: ApiT WalletName -> Text

Limits

passphraseMinLength :: Int

passphraseMaxLength :: Int

walletNameMinLength :: Int

walletNameMaxLength :: Int

Polymorphic Types

newtype ApiT a

Polymorphic wrapper type to put around primitive types and, 3rd party lib types to avoid defining orphan instances and/or, undesirable instances on primitive types. It helps to keep a nice separation of concerns between the API layer and other modules.

Constructors

ApiT

Fields

getApiT :: a

newtype ApiMnemonicT (sizes :: [Nat]) (purpose :: Symbol)

Representation of mnemonics at the API-level, using a polymorphic type in the lengths of mnemonics that are supported (and an underlying purpose). In practice, mnemonics correspond to passphrases or seeds, and although they're nice to manipulate as mnemonics from a user-perspective, carrying around a list of words doesn't really make sense for the business logic, which prefers manipulating scrubbed bytes directly.

data MyWallet
    { mnemonic :: ApiMnemonicT '[15,18,21,24] "root-seed"
    }

Note that the given Nats **have** to be valid mnemonic sizes, otherwise the underlying code won't even compile, with not-so-friendly error messages.

Also, the internal representation holds a [Text] which contains the list of mnemonic words that was parsed. This is only to be able to implement the ToJSON instances and roundtrip, which is a very dubious argument. In practice, we'll NEVER peek at the mnemonic, output them and whatnot.

Constructors

ApiMnemonicT (Passphrase purpose, [Text])

class MkApiMnemonic sizes purpose where

Methods

mkApiMnemonic :: [Text] -> Either MkApiMnemonicError (ApiMnemonicT sizes purpose)

newtype MkApiMnemonicError

Constructors

MkApiMnemonicError String

getApiMnemonicT :: ApiMnemonicT sizes purpose -> Passphrase purpose