Cardano.Wallet.Binary

Description

These are (partial) CBOR decoders for blocks and block headers. Note that we ignore most of the block's and header's content and only retrieve the pieces of information relevant to us, wallet (we do assume a trusted node and therefore, we needn't to care about verifying signatures and blocks themselves).

The format described in the decoders below are the one used in the Byron era of Cardano and will endure in the first stages of Shelley. They are also used by components like the Rust cardano-http-bridge.

Documentation

decodeBlock :: Decoder s Block

decodeBlockHeader :: Decoder s BlockHeader

Helpers

inspectNextToken :: Decoder s ()

Inspect the next token that has to be decoded and print it to the console as a trace. Useful for debugging Decoders. Example:

    myDecoder :: CBOR.Decoder s MyType
    myDecoder = do
        a <- CBOR.decodeWord64
        inspectNextToken
        [...]

decodeList :: forall s a. Decoder s a -> Decoder s [a]

Decode an list of known length. Very similar to decodeListIndef.

    myDecoder :: CBOR.Decoder s [MyType]
    myDecoder = decodeList decodeOne
      where
        decodeOne :: CBOR.Decoder s MyType

decodeListIndef :: forall s a. Decoder s a -> Decoder s [a]

Decode an arbitrary long list. CBOR introduce a "break" character to mark the end of the list, so we simply decode each item until we encounter a break character.

    myDecoder :: CBOR.Decoder s [MyType]
    myDecoder = decodeListIndef decodeOne
      where
        decodeOne :: CBOR.Decoder s MyType