Cardano.Wallet.BlockSyncer

Description

This module contains the ticking function that is responsible for invoking block acquisition functionality and executing it in periodic fashion.

Known limitations: the ticking function makes sure action is not executed on already consumed block, but does not check and handle block gaps (aka catching up).

Documentation

newtype BlockHeadersConsumed

Constructors

BlockHeadersConsumed [BlockHeader]

tickingFunction

Arguments

:: IO [Block]
a way to get a new block

-> (Block -> IO ())
action taken on a new block

-> Millisecond
tick time

-> BlockHeadersConsumed

-> IO ()