Cardano.Wallet.Network

Interface

data NetworkLayer m

Constructors

NetworkLayer

Fields

nextBlocks :: SlotId -> ExceptT ErrNetworkUnreachable m [Block]

Gets some blocks from the node. It will not necessarily return all the blocks that the node has, but will receive a reasonable-sized chunk. It will never return blocks from before the given slot. It may return an empty list if the node does not have any blocks from after the starting slot.

networkTip :: ExceptT ErrNetworkTip m (Hash "BlockHeader", BlockHeader)

Get the current network tip from the chain producer

postTx :: SignedTx -> ExceptT ErrPostTx m ()

Broadcast a transaction to the chain producer

Errors

newtype ErrNetworkUnreachable

Network is not reachable

Constructors

ErrNetworkUnreachable Text

data ErrNetworkTip

Error while trying to get the network tip

Constructors

ErrNetworkTipNetworkUnreachable ErrNetworkUnreachable

ErrNetworkTipNotFound

data ErrPostTx

Error while trying to send a transaction

Constructors

ErrPostTxNetworkUnreachable ErrNetworkUnreachable

ErrPostTxBadRequest Text

ErrPostTxProtocolFailure Text

Deprecated (to be removed)

data TickResult a

The result type of the element fetch function provided to tick.

Constructors

GotChunk !a
Have a result, and there may be more available.

Sleep
There is no result available now, so wait.

tick

Arguments

:: MonadIO m

=> (st -> m (TickResult a, st))
A way to get a new elements

-> (a -> m ())
Action to be taken on new elements

-> Millisecond
Tick time

-> st
Initial state.

-> m r

Repeatedly fetch data from a given source function, and call an action for each element retrieved.

If the data source indicates that it has no more data at present (Sleep), then sleep for the interval delay, and then try the fetch again.

listen :: NetworkLayer IO -> ([Block] -> IO ()) -> IO a

Retrieve blocks from a chain producer and execute some given action for each block.

drain :: NetworkLayer IO -> (SlotId -> [Block] -> IO ()) -> IO ()

Process all available blocks from the network layer with the given action. When the network layer reports no more blocks, stop.